[FLINK-13499][maprfs] Handle MapR dependency purely through reflection
This allows us to remove the MapR dependency from the module.
The MapR maven dependency has frequently caused issues.
